+++ CHANGELOG.md
@@ -11,9 +11,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - IDLabFunctions: added functions lookup and lookupWithDelimiter 
 - IDLabFunctionsTest: added unit tests for lookup functions 
+<<<<<<< HEAD
 - IDLabFunctionsTest: relocate csv files used by tests for lookup function
 - IDLabFunctions: silence stack trace
 - CSVW: filter out rows with a comment prefix
+- Output file path now gets checked before the mapping
 
 ### Fixed
 - Upgrade postgresql JDBC driver to 42.3.3